functions/requirements.txt
firebase_functions~=0.1.0
google-cloud-tasks>=2.16.1
google-cloud-logging>=3.9.0
